subset: move _add_method and _uniq_sort to subset/util submodule
