+++ Lib/fontTools/subset/__init__.py
@@ -8,6 +8,7 @@ from fontTools.ttLib.tables import otTables
 from fontTools.otlLib.maxContextCalc import maxCtxFont
 from fontTools.pens.basePen import NullPen
 from fontTools.misc.loggingTools import Timer
+from fontTools.subset.util import _add_method, _uniq_sort
 from fontTools.subset.cff import *
 import sys
 import struct
@@ -362,26 +363,6 @@ log.glyphs = MethodType(_log_glyphs, log)
 # main module's logger
 timer = Timer(logger=logging.getLogger("fontTools.subset.timer"))
 
-
-def _add_method(*clazzes):
-	"""Returns a decorator function that adds a new method to one or
-	more classes."""
-	def wrapper(method):
-		done = []
-		for clazz in clazzes:
-			if clazz in done: continue # Support multiple names of a clazz
-			done.append(clazz)
-			assert clazz.__name__ != 'DefaultTable', \
-					'Oops, table class not found.'
-			assert not hasattr(clazz, method.__name__), \
-					"Oops, class '%s' has method '%s'." % (clazz.__name__,
-									       method.__name__)
-			setattr(clazz, method.__name__, method)
-		return None
-	return wrapper
-
-def _uniq_sort(l):
-	return sorted(set(l))
 
 def _dict_subset(d, glyphs):
 	return {g:d[g] for g in glyphs}
